requirements.txt
mysql_connector_repackaged==0.3.1
pandas==1.5.1
streamlit==1.22.0
